+++ CHANGELOG.md
@@ -13,3 +13,4 @@
 - Add Rewards module
 - Add the param change proposal to the oracle module
 - Integrate Oracle on the app and add the gas less fees
+- Add wasmbinding queries for the oracle module